+++ cli/src/conf/conf_set.rs
@@ -51,7 +51,7 @@ pub(crate) async fn set_nexus_conf(
 
     json_output(&serde_json::to_value(&conf).unwrap())?;
 
-    match conf.save(&conf_path).await {
+    match conf.save_to_path(&conf_path).await {
         Ok(()) => {
             conf_handle.success();
             Ok(())