+++ cli/src/tool/tool_unregister.rs
@@ -6,7 +6,7 @@ use {
 /// Unregister a Tool based on the provided FQN.
 pub(crate) async fn unregister_tool(
     tool_fqn: ToolFqn,
-    owner_cap: sui::ObjectID,
+    owner_cap: Option<sui::ObjectID>,
     sui_gas_coin: Option<sui::ObjectID>,
     sui_gas_budget: u64,
     skip_confirmation: bool,
@@ -36,7 +36,13 @@ pub(crate) async fn unregister_tool(
     // Fetch reference gas price.
     let reference_gas_price = fetch_reference_gas_price(&sui).await?;
 
-    // Fetch the OwnerCap object.
+    // Use the provided or saved `owner_cap` object ID and fetch the object.
+    let Some(owner_cap) = owner_cap.or(conf.tools.get(&tool_fqn).map(|t| t.over_tool)) else {
+        return Err(NexusCliError::Any(anyhow!(
+            "No OwnerCap object ID found for tool '{tool_fqn}'."
+        )));
+    };
+
     let owner_cap = fetch_object_by_id(&sui, owner_cap).await?;
 
     // Craft a TX to unregister the tool.